Refactor: Extract interfaces for analyzer services
