+++ stats/analyzer/src/main/java/ru/practicum/ewm/controller/RecommendationsController.java
@@ -6,8 +6,8 @@ import io.grpc.stub.StreamObserver;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.devh.boot.grpc.server.service.GrpcService;
-import ru.practicum.ewm.service.EventSimilarityService;
-import ru.practicum.ewm.service.UserActionService;
+import ru.practicum.ewm.service.EventSimilarityServiceImpl;
+import ru.practicum.ewm.service.UserActionServiceImpl;
 import ru.practicum.ewm.stats.message.InteractionsCountRequestProto;
 import ru.practicum.ewm.stats.message.RecommendedEventProto;
 import ru.practicum.ewm.stats.message.SimilarEventsRequestProto;
@@ -20,15 +20,15 @@ import java.util.List;
 @RequiredArgsConstructor
 @Slf4j
 public class RecommendationsController extends RecommendationsControllerGrpc.RecommendationsControllerImplBase {
-    private final UserActionService userActionService;
-    private final EventSimilarityService eventSimilarityService;
+    private final UserActionServiceImpl userActionServiceImpl;
+    private final EventSimilarityServiceImpl eventSimilarityServiceImpl;
 
     @Override
     public void getInteractionsCount(InteractionsCountRequestProto request,
                                      StreamObserver<RecommendedEventProto> observer) {
         try {
             log.info("Received interactions weight request for eventIds: {}", request.getEventIdList());
-            List<RecommendedEventProto> events = userActionService.getTotalInteractionWeight(request.getEventIdList());
+            List<RecommendedEventProto> events = userActionServiceImpl.getTotalInteractionWeight(request.getEventIdList());
             events.forEach(event -> {
                 observer.onNext(event);
                 log.info("Response to interactions weight request: eventId: {}, rating: {}",
@@ -48,7 +48,7 @@ public class RecommendationsController extends RecommendationsControllerGrpc.Rec
         try {
             log.info("Received similar events request: userId: {}, eventId: {}, maxResults = {}",
                     request.getUserId(), request.getEventId(), request.getMaxResults());
-            List<RecommendedEventProto> events = eventSimilarityService.getSimilarEvents(
+            List<RecommendedEventProto> events = eventSimilarityServiceImpl.getSimilarEvents(
                     request.getUserId(),
                     request.getEventId(),
                     request.getMaxResults());
@@ -71,7 +71,7 @@ public class RecommendationsController extends RecommendationsControllerGrpc.Rec
         try {
             log.info("Received recommendations for user request: userId: {}, maxResults: {}",
                     request.getUserId(), request.getMaxResults());
-            List<RecommendedEventProto> events = eventSimilarityService.getRecommendationsForUser(request.getUserId(),
+            List<RecommendedEventProto> events = eventSimilarityServiceImpl.getRecommendationsForUser(request.getUserId(),
                     request.getMaxResults());
             events.forEach(event -> {
                 observer.onNext(event);
